requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.2.0
librosa==0.10.1
numpy==1.24.3
pydub==0.25.1
torch==2.1.0
msclap==1.3.3
soundfile==0.12.1
audioread==3.0.1 